# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Application API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the generated application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/suppliers/bulk-import": {
      "post": {
        "summary": "Bulk import suppliers from CSV",
        "tags": [
          "Suppliers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/suppliers/export": {
      "get": {
        "summary": "Export suppliers to CSV",
        "tags": [
          "Suppliers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/suppliers": {
      "get": {
        "summary": "Get all suppliers",
        "tags": [
          "Suppliers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Suppliers retrieved successfully"
          }
        }
      },
      "post": {
        "summary": "Create new supplier",
        "tags": [
          "Suppliers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Supplier created successfully"
          }
        }
      }
    },
    "/api/validations/{id}/resolve": {
      "put": {
        "summary": "Mark validation issue as resolved",
        "tags": [
          "Validations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}